refactored bot module
