+++ src/bot/conversation/timezone.py
@@ -4,7 +4,7 @@ from telegram import (
     KeyboardButton,
     ReplyKeyboardMarkup,
     ReplyKeyboardRemove,
-    Update,
+    Update
 )
 from telegram.ext import (
     CallbackContext,
@@ -12,16 +12,16 @@ from telegram.ext import (
     ContextTypes,
     ConversationHandler,
     MessageHandler,
-    filters,
+    filters
 )
 
-from constants import callback_data, states
-from constants.timezone import DEFAULT_TIMEZONE
+from bot.constants import callback_data, states
+from bot.constants.timezone import DEFAULT_TIMEZONE
+from bot.decorators.logger import async_error_logger
+from bot.get_timezone import get_timezone_from_location, get_timezone_from_text_message, set_timezone
+from bot.menu_button import COMMANDS, menu_button
 from core.config import settings
 from core.send_message import reply_message
-from decorators.logger import async_error_logger
-from get_timezone import get_timezone_from_location, get_timezone_from_text_message, set_timezone
-from menu_button import COMMANDS, menu_button
 
 ASK_FLAG = "ask_flag"
