+++ src/bot/decorators/safe_conversion.py
