+++ src/run_bot.py
@@ -1,4 +1,4 @@
-from bot import init_polling
+from bot.bot import init_polling
 
 if __name__ == "__main__":
     init_polling()
