+++ src/run_webhook_api.py
@@ -11,16 +11,17 @@ from starlette.middleware.authentication import AuthenticationMiddleware
 from starlette.requests import Request
 from starlette.responses import JSONResponse, Response
 from starlette.routing import Route
-from telegram import Bot, Update
+from telegram import Bot
 from telegram.error import TelegramError
 
 from api import context_models
-from bot import init_webhook
+from bot.api import telegram_routes
+from bot.bot import init_webhook
+from bot.service.bot_service import BotNotifierService
 from core.config import settings
 from core.logger import LOGGING_CONFIG, logger
 from middleware import TokenAuthBackend
 from service.api_client import APIService
-from service.bot_service import BotNotifierService
 
 
 async def start_bot() -> None:
@@ -59,20 +60,6 @@ async def healthcheck_api(request: Request) -> JSONResponse:
     except httpx.RequestError as error:
         logger.error("Failed to connect to database: %s", error)
     return JSONResponse(content=health.__dict__)
-
-
-async def telegram_webhook_api(request: Request) -> Response:
-    response = {}
-
-    try:
-        request_json = await request.json()
-        bot_app = request.app.state.bot_app
-        await bot_app.update_queue.put(Update.de_json(data=request_json, bot=bot_app.bot))
-    except JSONDecodeError as error:
-        logger.error("Got a JSONDecodeError: %s", error)
-        response = {"status_code": httpx.codes.BAD_REQUEST}
-
-    return Response(**response)
 
 
 async def request_to_context(request: Request, context):
@@ -141,13 +128,14 @@ exception_handlers = {HTTPException: http_exception}
 
 
 routes = [
-    Route("/telegramWebhookApi", telegram_webhook_api, methods=["POST"]),
     Route("/healthcheck", healthcheck_api, methods=["GET"]),
     Route("/bot/consultation/assign", consultation_assign, methods=["POST"]),
     Route("/bot/consultation/close", consultation_close, methods=["POST"]),
     Route("/bot/consultation/message", consultation_message, methods=["POST"]),
     Route("/bot/consultation/feedback", consultation_feedback, methods=["POST"]),
 ]
+
+routes.extend(telegram_routes)
 
 middleware = [Middleware(AuthenticationMiddleware, backend=TokenAuthBackend())]
